+++ examples/service/flex/plugin_version/pagination/plugin_version_pagination_example.go
@@ -1,0 +1,42 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v1 "github.com/RJPearson94/twilio-sdk-go/service/flex/v1"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var flexSession *v1.Flex
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	flexSession = twilio.NewWithCredentials(creds).Flex.V1
+}
+
+func main() {
+	paginator := flexSession.
+		Plugin("FPXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		Versions.
+		NewVersionsPaginator()
+
+	for paginator.Next() {
+		currentPage := paginator.CurrentPage()
+		log.Printf("%v plugin version(s) found on page %v", len(currentPage.Versions), currentPage.Meta.Page)
+	}
+
+	if paginator.Error() != nil {
+		log.Panicf("%s", paginator.Error())
+	}
+
+	log.Printf("Total number of plugin version(s) found: %v", len(paginator.Versions))
+}